+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 +   (improvement) Use `sass` instead of the deprecated `node-sass`.
 *   (bug) Always generate sourcemaps in Rollup.
+*   (improvement) Clear output directories before building.
 
 
 1.0.2